+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Plot of yield point determination added for ACT (like ATT and TBT)
+
 ## [v0.1.2] – 2024-02-20
 
 ### Added